[ASV-2083] Initializing the installation process inside install manager
